refactor(shared): quitar dependencias compartidas hacia sections
