+++ Web Ui/src/shared/styles/DialogBoxStyles.ts
